+++ src/default.rs
@@ -51,6 +51,7 @@ pub use crate::scalar::{
 
 pub use crate::simd::{
     // Custom gamma slices
+    gamma_to_linear_premultiply_rgba_slice,
     gamma_to_linear_slice,
     linear_to_gamma_slice,
     // f32 RGBA slices (alpha-preserving)
@@ -76,6 +77,7 @@ pub use crate::simd::{
     srgb_u16_to_linear_rgba_slice,
     // u16 slices
     srgb_u16_to_linear_slice,
+    unpremultiply_linear_to_gamma_rgba_slice,
     unpremultiply_linear_to_srgb_rgba_slice,
     unpremultiply_linear_to_srgb_u8_rgba_slice,
 };